fix: users tests
Some tests have a mock decorator in order to avoid real api calls and a skip decorator in the rest of them that did
not work, because the main task of this feature is to implement the pipeline. The upgrade and addition of new tests
will be done in a new issue.

Co-authored-by: Joao-Moura <jaummoura15@gmail.com>